// src/payments/payme/dto/create-transaction.dto.ts
import { TransactionMethods } from '../constants/transaction-methods';

export class CreateTransactionDto {
  method: TransactionMethods;
  params: {
    id: string;
    time: number;
    amount: number;
    account: {
      user_id: string;
      plan_id: string;
      selected_sport?: string;
    };
  };
}